Convert unit tests for freehand from Mocha to Jest to be consistent with the rest of the codebase.
