+++ src/tools/freehandMouseTool.test.js
@@ -1,0 +1,321 @@
+import FreehandMouseTool from './freehandMouseTool.js';
+import calculateFreehandStatistics from './shared/freehandUtils/calculateFreehandStatistics.js';
+import freeHandArea from './shared/freehandUtils/freeHandArea.js';
+import freeHandIntersect from './shared/freehandUtils/freeHandIntersect.js';
+import pointInFreehand from './shared/freehandUtils/pointInFreehand.js';
+import { FreehandHandleData } from './shared/freehandUtils/FreehandHandleData.js'
+
+jest.mock('./../manipulators/drawHandles.js');
+jest.mock('./../util/drawing.js');
+jest.mock('./../externalModules.js', () => ({
+  cornerstone: {
+    colors: {
+      getColormap: jest.fn().mockImplementation(() => {
+        return {
+          setNumberOfColors: jest.fn(),
+          setColor: jest.fn()
+        }
+      })
+    }
+  }
+}));
+jest.mock('./../stateManagement/toolState.js', () => ({
+  getToolState: jest.fn()
+}));
+
+describe('freeHandIntersect.js', function() {
+  let dataHandles;
+
+  beforeEach(() => {
+    // First 3 corners of a square.
+    dataHandles = [
+      new FreehandHandleData({
+          x: 0.0,
+          y: 0.0
+      }),
+      new FreehandHandleData({
+          x: 0.0,
+          y: 1.0
+      }),
+      new FreehandHandleData({
+          x: 1.0,
+          y: 1.0
+      })
+    ];
+  });
+
+  it('should return true if a new handle crosses any previous line', function() {
+    const candidateHandle = new FreehandHandleData({
+      x: -0.5,
+      y: 0.5
+    });
+    const doesIntersect = freeHandIntersect.newHandle(candidateHandle, dataHandles);
+
+    expect(doesIntersect).toBeTruthy();
+  });
+
+  it('should return false if a new handle doesn\'t cross any previous line', function() {
+    const candidateHandle = new FreehandHandleData({
+      x: 0.5,
+      y: 0.5
+    });
+    const doesIntersect = freeHandIntersect.newHandle(candidateHandle, dataHandles);
+
+    expect(doesIntersect).toBeDefined();
+    expect(doesIntersect).toBeFalsy();
+  });
+
+  it('should return true if the line created by finishing the polygon crosses any previous line', function() {
+    //Add an additional handle above the triangle such the line (handle3,handle0) crosses (handle1,handle2).
+    const handle3 = new FreehandHandleData({
+      x: 0.5,
+      y: 1.5
+    });
+
+    dataHandles.push(handle3);
+
+    const doesIntersect = freeHandIntersect.end(dataHandles);
+
+    expect(doesIntersect).toBeTruthy();
+  });
+
+  it('should return false if the line created by finishing the polygon doesn\'t cross any previous line', function() {
+    const doesIntersect = freeHandIntersect.end(dataHandles);
+
+    expect(doesIntersect).toBeDefined();
+    expect(doesIntersect).toBeFalsy();
+  });
+
+  it('should return true if one moves a polygon\'s vertex such that it causes an lines to cross', function() {
+    //add the 4th corner of the square first.
+    const handle3 = new FreehandHandleData({
+      x: 1.0,
+      y: 0.0
+    });
+    dataHandles.push(handle3);
+
+    //Move handle2 so that it crosses line (handle0,handle1)
+    const modifiedHandleId = 2;
+
+    dataHandles[modifiedHandleId].x = -0.5;
+    dataHandles[modifiedHandleId].y = 0.5;
+
+    const doesIntersect = freeHandIntersect.modify(dataHandles, modifiedHandleId);
+
+    expect(doesIntersect).toBeTruthy();
+  });
+
+  it('should return false if one moves a polygon\'s vertex a bit such that no lines cross', function() {
+    //add the 4th corner of the square first.
+    const handle3 = new FreehandHandleData({
+      x: 1.0,
+      y: 0.0
+    });
+    dataHandles.push(handle3);
+
+    //Move handle2 in a non disruptive way.
+    const modifiedHandleId = 2;
+
+    dataHandles[modifiedHandleId].x = 2.0;
+    dataHandles[modifiedHandleId].y = 2.0;
+
+    const doesIntersect = freeHandIntersect.modify(dataHandles, modifiedHandleId);
+
+    expect(doesIntersect).toBeDefined();
+    expect(doesIntersect).toBeFalsy();
+  });
+});
+
+describe('pointInFreehand.js', function() {
+  // Simple square
+  const dataHandles = [
+    new FreehandHandleData({
+        x: 0.0,
+        y: 0.0
+    }),
+    new FreehandHandleData({
+        x: 0.0,
+        y: 1.0
+    }),
+    new FreehandHandleData({
+        x: 1.0,
+        y: 1.0
+    }),
+    new FreehandHandleData({
+        x: 1.0,
+        y: 0.0
+    })
+  ];
+
+  it('should return true if the point is inside the polygon', function() {
+    const point = {
+      x: 0.5,
+      y: 0.5
+    };
+    const isInside = pointInFreehand(dataHandles, point);
+
+    expect(isInside).toBeTruthy();
+  });
+
+  it('should return false if the point is outside the object, to the right', function() {
+    const point = {
+      x: 2.0,
+      y: 0.5
+    };
+    const isInside = pointInFreehand(dataHandles, point);
+
+    expect(isInside).toBeDefined();
+    expect(isInside).toBeFalsy();
+  });
+
+  it('should return false if the point is outside the object, to the left', function() {
+    const point = {
+      x: -1.0,
+      y: 0.5
+    };
+    const isInside = pointInFreehand(dataHandles, point);
+
+    expect(isInside).toBeDefined();
+    expect(isInside).toBeFalsy();
+  });
+
+  it('should return false if on the line exactly', function() {
+    const point = {
+      x: 1.0,
+      y: 0.5
+    };
+    const isInside = pointInFreehand(dataHandles, point);
+
+    expect(isInside).toBeDefined();
+    expect(isInside).toBeFalsy();
+  });
+
+  it('should return false if the point is outside the object, above or bellow', function() {
+    const point1 = {
+      x: 0.5,
+      y: 2.0
+    };
+    const point2 = {
+      x: 0.5,
+      y: -1.0
+    };
+    const isInside1 = pointInFreehand(dataHandles, point1);
+    const isInside2 = pointInFreehand(dataHandles, point2);
+
+    expect(isInside1).toBeDefined();
+    expect(isInside1).toBeFalsy();
+    expect(isInside2).toBeDefined();
+    expect(isInside2).toBeFalsy();
+  });
+});
+
+
+describe('freeHandArea.js', function() {
+  // A unit square
+  const dataHandles = [
+    new FreehandHandleData({
+        x: 0.0,
+        y: 0.0
+    }),
+    new FreehandHandleData({
+        x: 0.0,
+        y: 1.0
+    }),
+    new FreehandHandleData({
+        x: 1.0,
+        y: 1.0
+    }),
+    new FreehandHandleData({
+        x: 1.0,
+        y: 0.0
+    })
+  ];
+
+  it('should return the area enclosed in dataHandles', function () {
+    const area = freeHandArea(dataHandles, false);
+
+    expect(area).toBeCloseTo(1.0, 5);
+  });
+
+  it('should scale if a scale parameter is given.', function () {
+    const area = freeHandArea(dataHandles, 10.0);
+
+    expect(area).toBeCloseTo(10.0, 5);
+  });
+
+});
+
+describe('calculateFreehandStatistics.js', function() {
+  // 'L' shape: 10x10 square with top right quadrant missing
+  const dataHandles = [
+    new FreehandHandleData({
+        x: 0.0,
+        y: 0.0
+    }),
+    new FreehandHandleData({
+        x: 0.0,
+        y: 10.0
+    }),
+    new FreehandHandleData({
+        x: 5.0,
+        y: 10.0
+    }),
+    new FreehandHandleData({
+        x: 5.0,
+        y: 5.0
+    }),
+    new FreehandHandleData({
+        x: 10.0,
+        y: 5.0
+    }),
+    new FreehandHandleData({
+        x: 10.0,
+        y: 0.0
+    })
+  ];
+
+  const boundingBox = {
+    top: 0.0,
+    height: 10.0,
+    left: 0.0,
+    width: 10.0
+  };
+
+  it('should not include pixels outside the polygon', function() {
+    // create 100 dimensional pixel array
+    const pixels = [];
+    pixels.length = 100;
+    pixels.fill(0.0);
+
+    // Set a pixel outside the polygon to a high value;
+    pixels[78] = 1000.0;
+
+    const statisticsObj = calculateFreehandStatistics(pixels, boundingBox, dataHandles);
+
+    expect(statisticsObj.mean).toBeCloseTo(0.0, 5);
+    expect(statisticsObj.variance).toBeCloseTo(0.0, 5);
+    expect(statisticsObj.stdDev).toBeCloseTo(0.0, 5);
+
+  });
+
+  it('should calculate that statistics of pixels within the polygon', function() {
+    // create 100 dimensional pixel array
+    const pixels = [];
+    pixels.length = 100;
+
+    // Set pixels within (0,0),(0,5),(10,5),(10,0) to 1
+    for (let i = 0; i < 50; i++) {
+      pixels[i] = 1.0;
+    }
+    // Set pixels within (0,5),(0,10),(10,10),(10,5) to 4
+    for (let i = 50; i < 100; i++) {
+      pixels[i] = 4.0;
+    }
+
+    const statisticsObj = calculateFreehandStatistics(pixels, boundingBox, dataHandles);
+
+    expect(statisticsObj.mean).toBeCloseTo(2.0, 5);
+    expect(statisticsObj.variance).toBeCloseTo(2.0, 5);
+    expect(statisticsObj.stdDev).toBeCloseTo(Math.sqrt(2.0), 5);
+  });
+});
